LFLT-480 Update dependencies of Leaflet stack components  - Changed CircleCI build image version  - Changed CircleCI Docker build environment version  - Updated Leaflet Stack Base BOM version to 2.0-r2208.1  - Fixed H2 incompatibility issues  - Bumped version to 1.12.0-dev
