+++ acceptance-tests/src/test/resources/application.yml
@@ -8,7 +8,7 @@ spring:
     defer-datasource-initialization: true
     generate-ddl: true
   datasource:
-    url: jdbc:h2:mem:acceptance
+    url: jdbc:h2:mem:acceptance;MODE=LEGACY
     username: sa
     password:
     driver-class-name: org.h2.Driver
