+++ pom.xml
@@ -6,12 +6,12 @@
 
     <artifactId>leaflet</artifactId>
     <packaging>pom</packaging>
-    <version>1.11.0-dev</version>
+    <version>1.12.0-dev</version>
 
     <parent>
         <groupId>hu.psprog.leaflet</groupId>
         <artifactId>leaflet-stack-base-bom</artifactId>
-        <version>1.0-r2110.1</version>
+        <version>2.0-r2208.1</version>
     </parent>
 
     <modules>
