+++ service/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <groupId>hu.psprog.leaflet</groupId>
         <artifactId>leaflet</artifactId>
-        <version>1.11.0-dev</version>
+        <version>1.12.0-dev</version>
     </parent>
     
     <artifactId>leaflet-service</artifactId>
